Fixed Spirit Shooting Mechanics v0.1
fixed some cooldown time

I just realised the issue with the shooting mechanic, thanks to the feedback from the Google form. The problem is that "the shooting isn't responsive it feels like the input is delayed." Initially, I thought it was due to game performance, but I discovered something else.

The problem was that I created the spirit body to be non-collidable. Although this isn't ideal with the current Spirit algorithm, when the shoot input is registered and the spirit is inside collision tilemaps, the bullet is actually coming out but collides with the inside tile, making it not visible, which causes a "BUG."

I have a plan to revamp how I designed the spirit. I will make the dog's current active state a singleton so that the spirit can access this information globally. For example, when the dog's current state is idle, the spirit will also be in an idle state but will move above the dog to ensure it doesn't overlap with anything before can_shoot is enabled. currently Spirit logic isn’t very robust.