Audio Manager v1
Contains, Sound Queue and SoundPool

what is sound pool?

Sound pool manages similar sounds such as footstep this script with play the sounds randomly to avoid repetitive.

what is sound queue?

Because i have a spirit that can shoot the bullet similutaniously.  Each bullet fired should make a sound. If the sprint shoot rapidly, there will be many sounds triggered in quick time and this will lead to sounds cut off.

With only one AudioStreamPlayer, the sound for each bullet would overlap and may not play correctly.

By duplicating the AudioStreamPlayer, you can play the sound for each bullet on a different instance, ensuring that each sound plays fully even if they overlap.

so TLDR without sound queue and just using one AudioStream each time shot is fired the sound is cutoff so that's why it is necessary for this algorithm